--- src/main/java/camp/nextstep/edu/calculator/Calculator.java
@@ -1,5 +1,0 @@
-package camp.nextstep.edu.calculator;
-
-public interface Calculator<T extends Number> {
-    T calculate(String expression);
-}
